use u128 for calculations

+++ programs/accountant/src/state/generic_accountant.rs
@@ -32,7 +32,7 @@ impl Accountant for GenericAccountant {
     }
 
     fn report(&self, profit: u64, _loss: u64) -> Result<(u64, u64)> {
-        let total_fees = self.performance_fee * profit / FEE_BPS;
+        let total_fees = (self.performance_fee as u128 * profit as u128 / FEE_BPS as u128) as u64;
         let total_refunds = 0;
         Ok((total_fees, total_refunds))
     }
